config(ticdc): add column-selectors for sink and split config file (#3573)
