+++ pkg/config/server_config.go
@@ -15,7 +15,6 @@ package config
 
 import (
 	"encoding/json"
-	"fmt"
 	"math"
 	"net"
 	"strings"
@@ -24,7 +23,6 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/pkg/config/outdated"
 	cerror "github.com/pingcap/ticdc/pkg/errors"
 	"github.com/pingcap/ticdc/pkg/security"
 	"go.uber.org/zap"
@@ -43,115 +41,6 @@ const (
 
 func init() {
 	StoreGlobalServerConfig(GetDefaultServerConfig())
-}
-
-var defaultReplicaConfig = &ReplicaConfig{
-	CaseSensitive:    true,
-	EnableOldValue:   true,
-	CheckGCSafePoint: true,
-	Filter: &FilterConfig{
-		Rules: []string{"*.*"},
-	},
-	Mounter: &MounterConfig{
-		WorkerNum: 16,
-	},
-	Sink: &SinkConfig{
-		Protocol: "default",
-	},
-	Cyclic: &CyclicConfig{
-		Enable: false,
-	},
-	Scheduler: &SchedulerConfig{
-		Tp:          "table-number",
-		PollingTime: -1,
-	},
-	Consistent: &ConsistentConfig{
-		Level:             "none",
-		MaxLogSize:        64,
-		FlushIntervalInMs: 1000,
-		Storage:           "",
-	},
-}
-
-// ReplicaConfig represents some addition replication config for a changefeed
-type ReplicaConfig replicaConfig
-
-type replicaConfig struct {
-	CaseSensitive    bool              `toml:"case-sensitive" json:"case-sensitive"`
-	EnableOldValue   bool              `toml:"enable-old-value" json:"enable-old-value"`
-	ForceReplicate   bool              `toml:"force-replicate" json:"force-replicate"`
-	CheckGCSafePoint bool              `toml:"check-gc-safe-point" json:"check-gc-safe-point"`
-	Filter           *FilterConfig     `toml:"filter" json:"filter"`
-	Mounter          *MounterConfig    `toml:"mounter" json:"mounter"`
-	Sink             *SinkConfig       `toml:"sink" json:"sink"`
-	Cyclic           *CyclicConfig     `toml:"cyclic-replication" json:"cyclic-replication"`
-	Scheduler        *SchedulerConfig  `toml:"scheduler" json:"scheduler"`
-	Consistent       *ConsistentConfig `toml:"consistent" json:"consistent"`
-}
-
-// Marshal returns the json marshal format of a ReplicationConfig
-func (c *ReplicaConfig) Marshal() (string, error) {
-	cfg, err := json.Marshal(c)
-	if err != nil {
-		return "", cerror.WrapError(cerror.ErrEncodeFailed, errors.Annotatef(err, "Unmarshal data: %v", c))
-	}
-	return string(cfg), nil
-}
-
-// Unmarshal unmarshals into *ReplicationConfig from json marshal byte slice
-func (c *ReplicaConfig) Unmarshal(data []byte) error {
-	return c.UnmarshalJSON(data)
-}
-
-// UnmarshalJSON unmarshals into *ReplicationConfig from json marshal byte slice
-func (c *ReplicaConfig) UnmarshalJSON(data []byte) error {
-	// The purpose of casting ReplicaConfig to replicaConfig is to avoid recursive calls UnmarshalJSON,
-	// resulting in stack overflow
-	r := (*replicaConfig)(c)
-	err := json.Unmarshal(data, &r)
-	if err != nil {
-		return cerror.WrapError(cerror.ErrDecodeFailed, err)
-	}
-	v1 := outdated.ReplicaConfigV1{}
-	err = v1.Unmarshal(data)
-	if err != nil {
-		return cerror.WrapError(cerror.ErrDecodeFailed, err)
-	}
-	r.fillFromV1(&v1)
-	return nil
-}
-
-// Clone clones a replication
-func (c *ReplicaConfig) Clone() *ReplicaConfig {
-	str, err := c.Marshal()
-	if err != nil {
-		log.Panic("failed to marshal replica config",
-			zap.Error(cerror.WrapError(cerror.ErrDecodeFailed, err)))
-	}
-	clone := new(ReplicaConfig)
-	err = clone.Unmarshal([]byte(str))
-	if err != nil {
-		log.Panic("failed to unmarshal replica config",
-			zap.Error(cerror.WrapError(cerror.ErrDecodeFailed, err)))
-	}
-	return clone
-}
-
-func (c *replicaConfig) fillFromV1(v1 *outdated.ReplicaConfigV1) {
-	if v1 == nil || v1.Sink == nil {
-		return
-	}
-	for _, dispatch := range v1.Sink.DispatchRules {
-		c.Sink.DispatchRules = append(c.Sink.DispatchRules, &DispatchRule{
-			Matcher:    []string{fmt.Sprintf("%s.%s", dispatch.Schema, dispatch.Name)},
-			Dispatcher: dispatch.Rule,
-		})
-	}
-}
-
-// GetDefaultReplicaConfig returns the default replica config
-func GetDefaultReplicaConfig() *ReplicaConfig {
-	return defaultReplicaConfig.Clone()
 }
 
 // SecurityConfig represents security config for server
